chore: move BoxDecorationPainter to individual file
