+++ kraken/lib/rendering.dart
@@ -6,6 +6,7 @@
  */
 
 export 'src/rendering/box_decoration.dart';
+export 'src/rendering/box_decoration_painter.dart';
 export 'src/dom/elements/canvas/canvas_painter.dart';
 export 'src/rendering/fallback_view.dart';
 export 'src/rendering/filter.dart';
